+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v8.4
 --------------------------
+- Auto redirect to the new ingredient when is renamed
 - Various bug fixes and improvements
 - For full details please refer to the CHANGELOG